Added page settings
