+++ designer/server/src/models/forms/editor-v2/pages.js
@@ -1,0 +1,62 @@
+import { getFormSpecificNavigation } from '~/src/models/forms/editor-v2/common.js'
+import {
+  editorv2Path,
+  formOverviewBackLink,
+  formOverviewPath
+} from '~/src/models/links.js'
+
+/**
+ * @param {FormMetadata} metadata
+ * @param {FormDefinition} definition
+ */
+export function pagesViewModel(metadata, definition) {
+  const formPath = formOverviewPath(metadata.slug)
+  const navigation = getFormSpecificNavigation(formPath, metadata, 'Editor')
+
+  const pageActions = [
+    {
+      text: 'Add new page',
+      href: editorv2Path(metadata.slug, 'page'),
+      classes: 'govuk-button--inverse'
+    }
+  ]
+
+  const extraPageActions = [
+    {
+      text: 'Re-order pages',
+      href: '/reorder',
+      classes: 'govuk-button--secondary'
+    },
+    {
+      text: 'Preview form',
+      href: '/preview',
+      classes: 'govuk-link govuk-link--inverse'
+    }
+  ]
+
+  if (definition.pages.length > 1) {
+    pageActions.push(...extraPageActions)
+  }
+
+  const pageListModel = {
+    ...definition,
+    formSlug: metadata.slug,
+    navigation,
+    pageHeading: {
+      text: 'Add and edit pages'
+    },
+    pageCaption: {
+      text: definition.name
+    },
+    pageActions
+  }
+
+  return {
+    ...pageListModel,
+    backLink: formOverviewBackLink(metadata.slug)
+  }
+}
+
+/**
+ * @import { FormMetadata, FormDefinition } from '@defra/forms-model'
+ */
