+++ designer/server/src/models/forms/editor-v2/question-details.js
@@ -1,0 +1,108 @@
+import { buildErrorList } from '~/src/common/helpers/build-error-details.js'
+import { insertValidationErrors, isCheckboxSelected } from '~/src/lib/utils.js'
+import {
+  GOVUK_LABEL__M,
+  SAVE_AND_CONTINUE,
+  baseModelFields,
+  getFormSpecificNavigation,
+  getPageNum
+} from '~/src/models/forms/editor-v2/common.js'
+import { formOverviewPath } from '~/src/models/links.js'
+
+/**
+ * @param {ValidationFailure<FormEditor> | undefined} validation
+ */
+function questionDetailsFields(validation) {
+  const formValues = validation?.formValues
+  return {
+    fields: {
+      question: {
+        name: 'question',
+        id: 'question',
+        label: {
+          text: 'Question',
+          classes: GOVUK_LABEL__M
+        },
+        value: formValues?.question,
+        ...insertValidationErrors(validation?.formErrors.question)
+      },
+      hintText: {
+        name: 'hintText',
+        id: 'hintText',
+        label: {
+          text: 'Hint text (optional)',
+          classes: GOVUK_LABEL__M
+        },
+        rows: 3,
+        value: formValues?.hintText,
+        ...insertValidationErrors(validation?.formErrors.hintText)
+      },
+      questionOptional: {
+        name: 'questionOptional',
+        id: 'questionOptional',
+        classes: 'govuk-checkboxes--small',
+        items: [
+          {
+            value: 'true',
+            text: 'Make this question optional',
+            checked: isCheckboxSelected(formValues?.questionOptional)
+          }
+        ]
+      },
+      shortDescription: {
+        name: 'shortDescription',
+        idPrefix: 'shortDescription',
+        label: {
+          text: 'Short description',
+          classes: GOVUK_LABEL__M
+        },
+        hint: {
+          text: "Enter a short description for this question like 'licence period'. Short descriptions are used in error messages and on the check your answers page."
+        },
+        value: formValues?.shortDescription,
+        ...insertValidationErrors(validation?.formErrors.shortDescription)
+      }
+    }
+  }
+}
+
+/**
+ * @param {FormMetadata} metadata
+ * @param {FormDefinition} definition
+ * @param {string} pageId
+ * @param {ValidationFailure<FormEditor>} [validation]
+ */
+export function questionDetailsViewModel(
+  metadata,
+  definition,
+  pageId,
+  validation
+) {
+  const pageTitle = metadata.title
+  const formPath = formOverviewPath(metadata.slug)
+  const navigation = getFormSpecificNavigation(formPath, metadata, 'Editor')
+  const { formErrors } = validation ?? {}
+
+  const pageNum = getPageNum(definition, pageId)
+
+  return {
+    ...baseModelFields(metadata.slug, pageTitle),
+    cardTitle: 'Question 1',
+    cardCaption: `Page ${pageNum}`,
+    navigation,
+    errorList: buildErrorList(formErrors, [
+      'question',
+      'shortDescription',
+      'hintText'
+    ]),
+    formErrors: validation?.formErrors,
+    formValues: validation?.formValues,
+    ...questionDetailsFields(validation),
+    buttonText: SAVE_AND_CONTINUE
+  }
+}
+
+/**
+ * @import { FormMetadata, FormDefinition, FormEditor } from '@defra/forms-model'
+ * @import { ValidationFailure } from '~/src/common/helpers/types.js'
+ */
