+++ designer/server/src/models/forms/editor-v2/question-type.js
@@ -1,0 +1,153 @@
+import { ComponentType } from '@defra/forms-model'
+
+import { buildErrorList } from '~/src/common/helpers/build-error-details.js'
+import { insertValidationErrors } from '~/src/lib/utils.js'
+import {
+  SAVE_AND_CONTINUE,
+  baseModelFields,
+  getFormSpecificNavigation,
+  getPageNum
+} from '~/src/models/forms/editor-v2/common.js'
+import { formOverviewPath } from '~/src/models/links.js'
+
+const questionTypeRadioItems = [
+  {
+    text: 'Written answer',
+    hint: {
+      text: 'A short or long answer as test or number'
+    },
+    value: 'written-answer-group'
+  },
+  {
+    text: 'Date',
+    hint: {
+      text: 'A day, month and year or month and year only'
+    },
+    value: 'date-group'
+  },
+  {
+    text: 'UK address',
+    hint: {
+      text: 'A street address, town or city and postcode'
+    },
+    value: ComponentType.UkAddressField
+  },
+  {
+    text: 'Phone number',
+    hint: {
+      text: 'A UK phone number, for example, 07700 900 982 or +44 808 157 0192'
+    },
+    value: ComponentType.TelephoneNumberField
+  },
+  {
+    text: 'Supporting evidence',
+    hint: {
+      text: 'A document, for example, DOC, PDF, CSV, Excel'
+    },
+    value: ComponentType.FileUploadField
+  },
+  {
+    text: 'Email address',
+    hint: {
+      text: 'An email address, for example, name@example.com'
+    },
+    value: ComponentType.EmailAddressField
+  },
+  {
+    divider: 'or'
+  },
+  {
+    text: 'A list of options that users can choose from',
+    value: ComponentType.SelectField
+  }
+]
+
+const writtenAnswerSubItems = [
+  {
+    text: 'Short answer (a single line)',
+    value: ComponentType.TextField
+  },
+  {
+    text: 'Long answer (more than a single line)',
+    value: ComponentType.MultilineTextField
+  },
+  { text: 'Numbers only', value: ComponentType.NumberField }
+]
+
+const dateSubItems = [
+  { text: 'Day, month and year', value: 'DatePartsField' },
+  { text: 'Month and year', value: 'MonthYearField' }
+]
+
+/**
+ * @param {FormMetadata} metadata
+ * @param {FormDefinition} definition
+ * @param {string} pageId
+ * @param {ValidationFailure<FormEditor>} [validation]
+ */
+export function questionTypeViewModel(
+  metadata,
+  definition,
+  pageId,
+  validation
+) {
+  const pageTitle = metadata.title
+  const formPath = formOverviewPath(metadata.slug)
+  const navigation = getFormSpecificNavigation(formPath, metadata, 'Editor')
+  const { formValues, formErrors } = validation ?? {}
+
+  const pageNum = getPageNum(definition, pageId)
+
+  return {
+    ...baseModelFields(metadata.slug, pageTitle),
+    navigation,
+    errorList: buildErrorList(formErrors, ['questionType']),
+    formErrors: validation?.formErrors,
+    formValues: validation?.formValues,
+    cardTitle: 'Question 1',
+    cardCaption: `Page ${pageNum}: question 1`,
+    fields: {
+      questionType: {
+        idPrefix: 'questionType',
+        name: 'questionType',
+        value: formValues?.questionType,
+        items: questionTypeRadioItems,
+        ...insertValidationErrors(validation?.formErrors.questionType)
+      },
+      writtenAnswerSub: {
+        id: 'writtenAnswerSub',
+        name: 'writtenAnswerSub',
+        idPrefix: 'writtenAnswerSub',
+        fieldset: {
+          legend: {
+            text: 'Type of written answer',
+            isPageHeading: false
+          }
+        },
+        items: writtenAnswerSubItems,
+        value: formValues?.writtenAnswerSub,
+        ...insertValidationErrors(validation?.formErrors.writtenAnswerSub)
+      },
+      dateSub: {
+        id: 'dateSub',
+        name: 'dateSub',
+        idPrefix: 'dateSub',
+        fieldset: {
+          legend: {
+            text: 'Type of date',
+            isPageHeading: false
+          }
+        },
+        items: dateSubItems,
+        value: formValues?.dateSub,
+        ...insertValidationErrors(validation?.formErrors.dateSub)
+      }
+    },
+    buttonText: SAVE_AND_CONTINUE
+  }
+}
+
+/**
+ * @import { FormMetadata, FormDefinition, FormEditor } from '@defra/forms-model'
+ * @import { ValidationFailure } from '~/src/common/helpers/types.js'
+ */
