+++ designer/server/src/routes/forms/editor-v2/questions.test.js
@@ -57,7 +57,7 @@ describe('Editor v2 questions routes', () => {
     name: 'Test form',
     pages: [
       {
-        id: '1',
+        id: 'f07fbbb1-268c-429b-bba5-5fc1f7353d7c',
         path: '/page-one',
         title: 'Page one',
         section: 'section',
@@ -100,7 +100,7 @@ describe('Editor v2 questions routes', () => {
     name: 'Test form',
     pages: [
       {
-        id: '1',
+        id: 'f07fbbb1-268c-429b-bba5-5fc1f7353d7c',
         path: '/page-one',
         title: 'Page one',
         section: 'section',
@@ -127,21 +127,17 @@ describe('Editor v2 questions routes', () => {
 
     const options = {
       method: 'get',
-      url: '/library/my-form-slug/editor-v2/page/1/questions',
+      url: '/library/my-form-slug/editor-v2/page/f07fbbb1-268c-429b-bba5-5fc1f7353d7c/questions',
       auth
     }
 
-    const { container, document } = await renderResponse(server, options)
+    const { container } = await renderResponse(server, options)
 
     const $mastheadHeading = container.getByText('Test form')
     const $cardTitle = container.getByText('Page 1 overview')
     const $cardHeading = container.getByText('Page 1')
-    const $questionNumbers = document.querySelectorAll(
-      'form .govuk-summary-list__key'
-    )
-    const $questionTitles = document.querySelectorAll(
-      'form .govuk-summary-list__value'
-    )
+    const $questionNumbers = container.getAllByRole('term')
+    const $questionTitles = container.getAllByRole('definition')
 
     const $actions = container.getAllByRole('button')
 
@@ -155,11 +151,12 @@ describe('Editor v2 questions routes', () => {
     expect($questionNumbers[0]).toHaveTextContent('Question 1')
     expect($questionNumbers[1]).toHaveTextContent('Question 2')
 
-    expect($questionTitles[0]).toHaveTextContent('This is your first question')
-    expect($questionTitles[1]).toHaveTextContent('This is your second question')
+    expect($questionTitles[1]).toHaveTextContent('This is your first question')
+    expect($questionTitles[3]).toHaveTextContent('This is your second question')
 
     expect($actions).toHaveLength(4)
     expect($actions[2]).toHaveTextContent('Add another question')
+    expect($actions[3]).toHaveTextContent('Save changes')
   })
 
   test('GET - should render no questions in the view', async () => {
@@ -170,21 +167,17 @@ describe('Editor v2 questions routes', () => {
 
     const options = {
       method: 'get',
-      url: '/library/my-form-slug/editor-v2/page/1/questions',
+      url: '/library/my-form-slug/editor-v2/page/f07fbbb1-268c-429b-bba5-5fc1f7353d7c/questions',
       auth
     }
 
-    const { container, document } = await renderResponse(server, options)
+    const { container } = await renderResponse(server, options)
 
     const $mastheadHeading = container.getByText('Test form')
     const $cardTitle = container.getByText('Page 1 overview')
     const $cardHeading = container.getByText('Page 1')
-    const $questionNumbers = document.querySelectorAll(
-      'form .govuk-summary-list__key'
-    )
-    const $questionTitles = document.querySelectorAll(
-      'form .govuk-summary-list__value'
-    )
+    const $questionNumbers = container.getAllByRole('term')
+    const $questionTitles = container.getAllByRole('definition')
 
     expect($mastheadHeading).toHaveTextContent('Test form')
     expect($mastheadHeading).toHaveClass('govuk-heading-xl')
@@ -194,7 +187,7 @@ describe('Editor v2 questions routes', () => {
     expect($cardHeading).toHaveClass('govuk-heading-l')
 
     expect($questionNumbers).toHaveLength(1)
-    expect($questionTitles).toHaveLength(1)
+    expect($questionTitles).toHaveLength(2)
     expect($questionNumbers[0]).toHaveTextContent('')
     expect($questionTitles[0]).toHaveTextContent('No questions')
   })
